requirements.txt
@@ -0,0 +1,104 @@
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
dateparser==1.2.0
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.2
Flask-Mail==0.10.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
frozenlist==1.4.1
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gspread==6.1.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
kiwisolver==1.4.7
langdetect==1.0.9
Mako==1.3.6
MarkupSafe==3.0.2
matplotlib==3.9.2
multidict==6.1.0
numpy==2.1.2
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.28.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==10.4.0
priority==2.0.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.0
pyparsing==3.1.4
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
Quart==0.19.9
quart-cors==0.7.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tqdm==4.67.1
twilio==9.3.7
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.3
websockets==13.1
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
yarl==1.14.0